Fixup jellarr.service lifecycle
Currently I find when I deploy Jellarr, nothing happens unless I
manually [re]start the service. I guess eventually if I waited for the
timer it would work, but that's not in line with my expectation for
deploying Nix services.

Add it to the wants for multi-user.target so that it is started
automatically as soon as it exists. Also, set RemainAfterExit so that
systemd considers the service to be "alive". This means that when the
configuration is changed, the NixOS configuration switch logic will
restart it, immediately deploying the updated configuration.